refactor: move misc API responces
